add some stuff in libftmath

+++ libft/src/libft_math/ft_abs.c
@@ -6,13 +6,13 @@
 /*   By: jpriou <jpriou@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/11 15:29:24 by jpriou            #+#    #+#             */
-/*   Updated: 2017/11/11 15:30:06 by jpriou           ###   ########.fr       */
+/*   Updated: 2018/08/02 17:30:24 by jpriou           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_abs(int c)
+double		ft_abs(double c)
 {
 	return ((c < 0) ? -c : c);
 }
