+++ libft/src/libft_math/ft_logb.c
@@ -6,20 +6,20 @@
 /*   By: jpriou <jpriou@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/12 07:47:18 by jpriou            #+#    #+#             */
-/*   Updated: 2017/11/12 07:59:15 by jpriou           ###   ########.fr       */
+/*   Updated: 2018/08/02 17:31:23 by jpriou           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static int		ft_log_pos(int n, int len)
+static double		ft_log_pos(double n, int len)
 {
 	if (n == 0)
 		return (0);
 	return (1 + ft_log_pos(n / len, len));
 }
 
-int				ft_logb(int n, int lengthbase)
+double				ft_logb(double n, int lengthbase)
 {
 	if (lengthbase <= 1)
 		return (-1);
