+++ libft/src/libft_math/ft_min.c
@@ -6,13 +6,13 @@
 /*   By: jpriou <jpriou@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/12 07:33:40 by jpriou            #+#    #+#             */
-/*   Updated: 2017/11/12 07:36:44 by jpriou           ###   ########.fr       */
+/*   Updated: 2018/08/02 17:31:41 by jpriou           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_min(int a, int b)
+double		ft_min(double a, double b)
 {
 	return ((a < b) ? a : b);
 }
